made SearchFetchEvents component to clean up events page
